refactor(execution): add a function for calc_message_l1_resources
